SearchBar - material3 (#101)
* Added the new component from material3

* Removed unnecessary files added to git

* Removed the extra import
