--- .idea/detekt.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="DetektPluginSettings">
-    <option name="configurationFilePaths">
-      <set>
-        <option value="$PROJECT_DIR$/app/config/detekt/detekt.yaml" />
-      </set>
-    </option>
-    <option name="enableFormatting" value="true" />
-  </component>
-</project>
