+++ ui/search/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
     implementation(libs.androidx.lifecycle.runtimeCompose)
     implementation(libs.androidx.hilt.navigation.compose)
     implementation(libs.androidx.navigation.compose)
+    implementation(libs.androidx.compose.material3)
     implementation(libs.accompanist.systemuicontroller)
     implementation(libs.hilt.android)
     kapt(libs.hilt.compiler)
